build: run post-install scripts on post-upgrade
When using apk, a user and a group defined using USERID in a package
Makefile doesn't get created when it's defined in a new version of a
package and that version of a package is installed over an existing
version (i.e. upgraded) without USERID defined. It only works when a new
package with USERID defined is installed due to post-install scripts not
running on upgrade.

Run post-install scripts that create the user and group on post-upgrade.

Link: https://wiki.alpinelinux.org/wiki/Creating_an_Alpine_package
Fixes: #18527
Suggested-by: Thomas Richard <thomas.richard@bootlin.com>
Signed-off-by: George Sapkin <george@sapk.in>